feat: Add OQTOPUS experiments utilities package
- Introduced a comprehensive utilities package for OQTOPUS experiments, including modules for data management, display, statistical processing, and OQTOPUS-specific utilities.
- Implemented functions for creating and saving experiment data structures, generating metadata, and loading/saving JSON files.
- Added display utilities for formatting and presenting experiment results using Rich tables with a fallback to plain text.
- Included statistical functions for probability calculations, parameter estimation, and fitting quality assessment.
- Consolidated conversion utilities for handling measurement counts and validating OQTOPUS results.
- Enhanced modularity and reusability across different experiment types by reducing code duplication.
