+++ src/oqtopus_experiments/cli/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-CLI Framework for QuantumLib Experiments
+CLI Framework for OQTOPUS Experiments Experiments
 """
 
 from .base_cli import BaseExperimentCLI, DeviceType, ExperimentBackend
