+++ src/oqtopus_experiments/cli/ramsey.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-T2 Echo CLI - QuantumLib T2 Echo Experiment (Hahn Echo/CPMG)
+Ramsey CLI - OQTOPUS Experiments Ramsey Oscillation Experiment
 """
 
 from typing import Annotated, Any
@@ -8,7 +8,7 @@ from typing import Annotated, Any
 import numpy as np
 import typer
 
-from quantumlib.cli.base_cli import (
+from oqtopus_experiments.cli.base_cli import (
     BaseExperimentCLI,
     CommonBackendOption,
     CommonDevicesOption,
@@ -22,106 +22,100 @@ from quantumlib.cli.base_cli import (
     DeviceType,
     ExperimentBackend,
 )
-from quantumlib.experiments.t2_echo.t2_echo_experiment import T2EchoExperiment
+from oqtopus_experiments.experiments.ramsey.ramsey_experiment import RamseyExperiment
 
 
-class T2EchoExperimentCLI(BaseExperimentCLI):
+class RamseyExperimentCLI(BaseExperimentCLI):
     """
-    T2 Echo experiment dedicated CLI (using QuantumLib integrated framework)
+    Ramsey experiment dedicated CLI (using OQTOPUS Experiments integrated framework)
     """
 
     def __init__(self):
         super().__init__(
-            experiment_name="T2Echo",
-            help_text="QuantumLib T2 Echo Experiment (Hahn Echo/CPMG)",
+            experiment_name="Ramsey",
+            help_text="OQTOPUS Experiments Ramsey Oscillation Experiment",
         )
 
     def get_experiment_class(self):
-        """Returns the T2 Echo experiment class"""
-        return T2EchoExperiment
+        """Returns the Ramsey experiment class"""
+        return RamseyExperiment
 
     def get_experiment_specific_options(self) -> dict[str, Any]:
-        """T2 Echo experiment specific options"""
+        """Ramsey experiment specific options"""
         return {
             "delay_points": 51,
-            "max_delay": 500000,  # 500μs
-            "echo_type": "hahn",
-            "num_echoes": 1,
+            "max_delay": 200000,
+            "detuning": 0.0,
             "enable_fitting": True,
         }
 
     def create_experiment_config_display(self, **kwargs) -> str:
-        """T2 Echo experiment configuration display"""
+        """Ramsey experiment configuration display"""
         devices = kwargs.get("devices", ["qulacs"])
         backend = kwargs.get("backend", "local_simulator")
         shots = kwargs.get("shots", 1000)
         parallel = kwargs.get("parallel", 4)
         delay_points = kwargs.get("delay_points", 51)
-        max_delay = kwargs.get("max_delay", 500000)
-        echo_type = kwargs.get("echo_type", "hahn")
-        num_echoes = kwargs.get("num_echoes", 1)
+        max_delay = kwargs.get("max_delay", 200000)
+        detuning = kwargs.get("detuning", 0.0)
         enable_fitting = kwargs.get("enable_fitting", True)
         return (
-            f"QuantumLib T2 Echo Experiment\\n"
+            f"OQTOPUS Experiments Ramsey Oscillation Experiment\\n"
             f"Devices: {', '.join(devices)}\\n"
             f"Backend: {backend}\\n"
             f"Shots: {shots:,} per delay | Points: {delay_points}\\n"
-            f"Max Delay: {max_delay / 1000:.1f} μs | Echo: {echo_type.upper()}\\n"
-            f"Echo Count: {num_echoes} | Fitting: {'Enabled' if enable_fitting else 'Disabled'}\\n"
+            f"Max Delay: {max_delay / 1000:.1f} μs | Detuning: {detuning} MHz\\n"
+            f"Fitting: {'Enabled' if enable_fitting else 'Disabled'}\\n"
             f"Parallel: {parallel} threads\\n"
             f"Total measurements: {delay_points} delay points"
         )
 
     def generate_circuits(
-        self, experiment_instance: T2EchoExperiment, **kwargs
+        self, experiment_instance: RamseyExperiment, **kwargs
     ) -> tuple[list[Any], dict]:
-        """T2 Echo circuit generation"""
+        """Ramsey circuit generation"""
         delay_points = kwargs.get("delay_points", 51)
-        max_delay = kwargs.get("max_delay", 500000)
-        echo_type = kwargs.get("echo_type", "hahn")
-        num_echoes = kwargs.get("num_echoes", 1)
+        max_delay = kwargs.get("max_delay", 200000)
+        detuning = kwargs.get("detuning", 0.0)
 
         # Default delay time configuration
-        delay_times = np.logspace(np.log10(100), np.log10(500 * 1000), num=51)
+        delay_times = np.logspace(np.log10(50), np.log10(200 * 1000), num=51)
 
         circuits = experiment_instance.create_circuits(
             delay_points=delay_points,
             max_delay=max_delay,
-            echo_type=echo_type,
-            num_echoes=num_echoes,
+            detuning=detuning,
             delay_times=delay_times,
         )
 
         self.console.print(
             f"   Delay range: {delay_points} points from {delay_times[0]:.1f} to {delay_times[-1]:.1f} ns"
         )
-        self.console.print(f"   Echo type: {echo_type.upper()} (echoes={num_echoes})")
+        self.console.print(f"   Detuning: {detuning} MHz")
 
         return circuits, {
             "delay_times": delay_times,
             "max_delay": max_delay,
             "delay_points": delay_points,
-            "echo_type": echo_type,
-            "num_echoes": num_echoes,
+            "detuning": detuning,
         }
 
     def process_results(
         self,
-        experiment_instance: T2EchoExperiment,
+        experiment_instance: RamseyExperiment,
         raw_results: dict,
         circuits: list,
         metadata: Any,
         **kwargs,
     ) -> dict:
-        """T2 Echo result processing"""
+        """Ramsey result processing"""
         delay_times = metadata["delay_times"]
         max_delay = metadata["max_delay"]
         delay_points = metadata["delay_points"]
-        echo_type = metadata["echo_type"]
-        num_echoes = metadata["num_echoes"]
+        detuning = metadata["detuning"]
         enable_fitting = kwargs.get("enable_fitting", True)
 
-        self.console.print("   → Analyzing T2 Echo decay...")
+        self.console.print("   → Analyzing Ramsey oscillation...")
         analysis = experiment_instance.analyze_results(
             raw_results, enable_fitting=enable_fitting
         )
@@ -131,8 +125,7 @@ class T2EchoExperimentCLI(BaseExperimentCLI):
             "delay_times": delay_times.tolist(),
             "max_delay": max_delay,
             "delay_points": delay_points,
-            "echo_type": echo_type,
-            "num_echoes": num_echoes,
+            "detuning": detuning,
             "enable_fitting": enable_fitting,
         }
 
@@ -140,7 +133,7 @@ class T2EchoExperimentCLI(BaseExperimentCLI):
             "delay_times": delay_times,
             "device_results": analysis["device_results"],
             "analysis": analysis,
-            "method": "t2_echo_quantumlib_framework",
+            "method": "ramsey_quantumlib_framework",
         }
 
     def run(
@@ -159,19 +152,16 @@ class T2EchoExperimentCLI(BaseExperimentCLI):
         ] = 51,
         max_delay: Annotated[
             int, typer.Option(help="Maximum delay time [ns]")
-        ] = 500000,
-        echo_type: Annotated[
-            str, typer.Option(help="Echo pulse sequence type (hahn/cpmg)")
-        ] = "hahn",
-        num_echoes: Annotated[
-            int, typer.Option(help="Number of echo pulses")
-        ] = 1,
+        ] = 200000,
+        detuning: Annotated[
+            float, typer.Option(help="Frequency detuning [MHz]")
+        ] = 0.0,
         enable_fitting: Annotated[
-            bool, typer.Option(help="Enable T2 echo time fitting")
+            bool, typer.Option(help="Enable T2*/detuning parameter fitting")
         ] = True,
     ):
         """
-        Run T2 Echo experiment (Hahn Echo/CPMG)
+        Run Ramsey oscillation experiment
         """
         # Call framework's common execution logic
         self._execute_experiment(
@@ -184,11 +174,10 @@ class T2EchoExperimentCLI(BaseExperimentCLI):
             no_plot=no_plot,
             show_plot=show_plot,
             verbose=verbose,
-            delay_points=delay_points,  # T2 Echo specific option
-            max_delay=max_delay,  # T2 Echo specific option
-            echo_type=echo_type,  # T2 Echo specific option
-            num_echoes=num_echoes,  # T2 Echo specific option
-            enable_fitting=enable_fitting,  # T2 Echo specific option
+            delay_points=delay_points,  # Ramsey specific option
+            max_delay=max_delay,  # Ramsey specific option
+            detuning=detuning,  # Ramsey specific option
+            enable_fitting=enable_fitting,  # Ramsey specific option
         )
 
     def main(self):
@@ -197,8 +186,8 @@ class T2EchoExperimentCLI(BaseExperimentCLI):
 
 
 def main():
-    """Main entry point for quantumlib-t2-echo CLI command"""
-    cli = T2EchoExperimentCLI()
+    """Main entry point for quantumlib-ramsey CLI command"""
+    cli = RamseyExperimentCLI()
     cli.start()
