+++ src/oqtopus_experiments/cli/t2_echo.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-Ramsey CLI - QuantumLib Ramsey Oscillation Experiment
+T2 Echo CLI - OQTOPUS Experiments T2 Echo Experiment (Hahn Echo/CPMG)
 """
 
 from typing import Annotated, Any
@@ -8,7 +8,7 @@ from typing import Annotated, Any
 import numpy as np
 import typer
 
-from quantumlib.cli.base_cli import (
+from oqtopus_experiments.cli.base_cli import (
     BaseExperimentCLI,
     CommonBackendOption,
     CommonDevicesOption,
@@ -22,100 +22,106 @@ from quantumlib.cli.base_cli import (
     DeviceType,
     ExperimentBackend,
 )
-from quantumlib.experiments.ramsey.ramsey_experiment import RamseyExperiment
+from oqtopus_experiments.experiments.t2_echo.t2_echo_experiment import T2EchoExperiment
 
 
-class RamseyExperimentCLI(BaseExperimentCLI):
+class T2EchoExperimentCLI(BaseExperimentCLI):
     """
-    Ramsey experiment dedicated CLI (using QuantumLib integrated framework)
+    T2 Echo experiment dedicated CLI (using OQTOPUS Experiments integrated framework)
     """
 
     def __init__(self):
         super().__init__(
-            experiment_name="Ramsey",
-            help_text="QuantumLib Ramsey Oscillation Experiment",
+            experiment_name="T2Echo",
+            help_text="OQTOPUS Experiments T2 Echo Experiment (Hahn Echo/CPMG)",
         )
 
     def get_experiment_class(self):
-        """Returns the Ramsey experiment class"""
-        return RamseyExperiment
+        """Returns the T2 Echo experiment class"""
+        return T2EchoExperiment
 
     def get_experiment_specific_options(self) -> dict[str, Any]:
-        """Ramsey experiment specific options"""
+        """T2 Echo experiment specific options"""
         return {
             "delay_points": 51,
-            "max_delay": 200000,
-            "detuning": 0.0,
+            "max_delay": 500000,  # 500μs
+            "echo_type": "hahn",
+            "num_echoes": 1,
             "enable_fitting": True,
         }
 
     def create_experiment_config_display(self, **kwargs) -> str:
-        """Ramsey experiment configuration display"""
+        """T2 Echo experiment configuration display"""
         devices = kwargs.get("devices", ["qulacs"])
         backend = kwargs.get("backend", "local_simulator")
         shots = kwargs.get("shots", 1000)
         parallel = kwargs.get("parallel", 4)
         delay_points = kwargs.get("delay_points", 51)
-        max_delay = kwargs.get("max_delay", 200000)
-        detuning = kwargs.get("detuning", 0.0)
+        max_delay = kwargs.get("max_delay", 500000)
+        echo_type = kwargs.get("echo_type", "hahn")
+        num_echoes = kwargs.get("num_echoes", 1)
         enable_fitting = kwargs.get("enable_fitting", True)
         return (
-            f"QuantumLib Ramsey Oscillation Experiment\\n"
+            f"OQTOPUS Experiments T2 Echo Experiment\\n"
             f"Devices: {', '.join(devices)}\\n"
             f"Backend: {backend}\\n"
             f"Shots: {shots:,} per delay | Points: {delay_points}\\n"
-            f"Max Delay: {max_delay / 1000:.1f} μs | Detuning: {detuning} MHz\\n"
-            f"Fitting: {'Enabled' if enable_fitting else 'Disabled'}\\n"
+            f"Max Delay: {max_delay / 1000:.1f} μs | Echo: {echo_type.upper()}\\n"
+            f"Echo Count: {num_echoes} | Fitting: {'Enabled' if enable_fitting else 'Disabled'}\\n"
             f"Parallel: {parallel} threads\\n"
             f"Total measurements: {delay_points} delay points"
         )
 
     def generate_circuits(
-        self, experiment_instance: RamseyExperiment, **kwargs
+        self, experiment_instance: T2EchoExperiment, **kwargs
     ) -> tuple[list[Any], dict]:
-        """Ramsey circuit generation"""
+        """T2 Echo circuit generation"""
         delay_points = kwargs.get("delay_points", 51)
-        max_delay = kwargs.get("max_delay", 200000)
-        detuning = kwargs.get("detuning", 0.0)
+        max_delay = kwargs.get("max_delay", 500000)
+        echo_type = kwargs.get("echo_type", "hahn")
+        num_echoes = kwargs.get("num_echoes", 1)
 
         # Default delay time configuration
-        delay_times = np.logspace(np.log10(50), np.log10(200 * 1000), num=51)
+        delay_times = np.logspace(np.log10(100), np.log10(500 * 1000), num=51)
 
         circuits = experiment_instance.create_circuits(
             delay_points=delay_points,
             max_delay=max_delay,
-            detuning=detuning,
+            echo_type=echo_type,
+            num_echoes=num_echoes,
             delay_times=delay_times,
         )
 
         self.console.print(
             f"   Delay range: {delay_points} points from {delay_times[0]:.1f} to {delay_times[-1]:.1f} ns"
         )
-        self.console.print(f"   Detuning: {detuning} MHz")
+        self.console.print(f"   Echo type: {echo_type.upper()} (echoes={num_echoes})")
 
         return circuits, {
             "delay_times": delay_times,
             "max_delay": max_delay,
             "delay_points": delay_points,
-            "detuning": detuning,
+            "echo_type": echo_type,
+            "num_echoes": num_echoes,
         }
 
     def process_results(
         self,
-        experiment_instance: RamseyExperiment,
+        experiment_instance: T2EchoExperiment,
         raw_results: dict,
         circuits: list,
         metadata: Any,
         **kwargs,
     ) -> dict:
-        """Ramsey result processing"""
+        """T2 Echo result processing"""
         delay_times = metadata["delay_times"]
         max_delay = metadata["max_delay"]
         delay_points = metadata["delay_points"]
-        detuning = metadata["detuning"]
+        echo_type = metadata["echo_type"]
+        num_echoes = metadata["num_echoes"]
         enable_fitting = kwargs.get("enable_fitting", True)
 
-        self.console.print("   → Analyzing Ramsey oscillation...")
+        self.console.print("   → Analyzing T2 Echo decay...")
         analysis = experiment_instance.analyze_results(
             raw_results, enable_fitting=enable_fitting
         )
@@ -125,7 +131,8 @@ class RamseyExperimentCLI(BaseExperimentCLI):
             "delay_times": delay_times.tolist(),
             "max_delay": max_delay,
             "delay_points": delay_points,
-            "detuning": detuning,
+            "echo_type": echo_type,
+            "num_echoes": num_echoes,
             "enable_fitting": enable_fitting,
         }
 
@@ -133,7 +140,7 @@ class RamseyExperimentCLI(BaseExperimentCLI):
             "delay_times": delay_times,
             "device_results": analysis["device_results"],
             "analysis": analysis,
-            "method": "ramsey_quantumlib_framework",
+            "method": "t2_echo_quantumlib_framework",
         }
 
     def run(
@@ -152,16 +159,19 @@ class RamseyExperimentCLI(BaseExperimentCLI):
         ] = 51,
         max_delay: Annotated[
             int, typer.Option(help="Maximum delay time [ns]")
-        ] = 200000,
-        detuning: Annotated[
-            float, typer.Option(help="Frequency detuning [MHz]")
-        ] = 0.0,
+        ] = 500000,
+        echo_type: Annotated[
+            str, typer.Option(help="Echo pulse sequence type (hahn/cpmg)")
+        ] = "hahn",
+        num_echoes: Annotated[
+            int, typer.Option(help="Number of echo pulses")
+        ] = 1,
         enable_fitting: Annotated[
-            bool, typer.Option(help="Enable T2*/detuning parameter fitting")
+            bool, typer.Option(help="Enable T2 echo time fitting")
         ] = True,
     ):
         """
-        Run Ramsey oscillation experiment
+        Run T2 Echo experiment (Hahn Echo/CPMG)
         """
         # Call framework's common execution logic
         self._execute_experiment(
@@ -174,10 +184,11 @@ class RamseyExperimentCLI(BaseExperimentCLI):
             no_plot=no_plot,
             show_plot=show_plot,
             verbose=verbose,
-            delay_points=delay_points,  # Ramsey specific option
-            max_delay=max_delay,  # Ramsey specific option
-            detuning=detuning,  # Ramsey specific option
-            enable_fitting=enable_fitting,  # Ramsey specific option
+            delay_points=delay_points,  # T2 Echo specific option
+            max_delay=max_delay,  # T2 Echo specific option
+            echo_type=echo_type,  # T2 Echo specific option
+            num_echoes=num_echoes,  # T2 Echo specific option
+            enable_fitting=enable_fitting,  # T2 Echo specific option
         )
 
     def main(self):
@@ -186,8 +197,8 @@ class RamseyExperimentCLI(BaseExperimentCLI):
 
 
 def main():
-    """Main entry point for quantumlib-ramsey CLI command"""
-    cli = RamseyExperimentCLI()
+    """Main entry point for quantumlib-t2-echo CLI command"""
+    cli = T2EchoExperimentCLI()
     cli.start()
