+++ src/oqtopus_experiments/templates/chsh_examples.py
